add Assigned to filter for the tasks list
